refactor: moved the core logic to Utils for a clean BST class
